Changed generation PDF in randomized regression
In ANNZ::getRndMethodBestPDF(), the PDF was constructed by selecting
the MLMs with the best performance in terms of scatter and outlier
fractions. This has been changed to take into account the combination
of scatter and bias instead.

Also added the user-parameters: “max_sigma68_PDF”, “max_bias_PDF” and
“max_frac68_PDF”. These allow the user to define the maximal value of
the scatter, bias and combined outlier-fraction metrics of MLMs which
are included in PDFs in randomized regression.