ANNZ 2.0.5
Modified the condition on the MLM error estimator which is propagated
to PDFs in randomized regression. Previously if an error was undefined,
indicating a problem, the code would stop. Now the MLM is ignored and
the code continues. The change is needed as sometimes there is a valid
reason for the error to be undefined.